% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_list.R
\name{apply_filter_list}
\alias{apply_filter_list}
\title{Applies filtering threshold on the matrix of interaction metrics}
\usage{
apply_filter_list(calc_metrics, filter_list = NULL)
}
\arguments{
\item{calc_metrics}{A list of calculated metrics obtained by applying
metric_functions onto a data matrix.}

\item{filter_list}{A list of threshold values to be applied on the calc_metrics.
All the values with abslolute value smaller than the threshold is set to 0.
Default: NULL}
}
\description{
Applies filtering threshold on the matrix of interaction metrics
}
\examples{
mat = matrix(sample(100, 300, replace = TRUE), 30, 10,
		   dimnames = list(paste0("c", 1:30), letters[1:10]))
mat[, 2] = 2 * mat[,1] + mat[,2]
mat[runif(300) < 0.5] = 0


metric_functions = list(CoDa = cor_coda,
                        cooccurrence = cooccurrence,
                        MI = bcmi,
                        Spearman = function(x) {cor_sigs(x, method = "spearman")})

filter_threshold = list(MI = 0.2)

#'
metrics_out = lapply(metric_functions, function(x) x(mat))


filtered_metrics_out = apply_filter_list(calc_metrics = metrics_out,
                                         filter_list = filter_threshold)

}
