% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_coda_novel.R
\name{cor_coda_novel}
\alias{cor_coda_novel}
\title{Pairwise signature correlations from symmetric pivot coordinates}
\usage{
cor_coda_novel(x, p.val = 0.05, p.adjust = TRUE, ...)
}
\arguments{
\item{x}{Data.frame or a matrix of signature intensities(or counts).
Signatures need to correspond to columns.}

\item{p.val}{All the correlations with >=p.val are set to 0.}

\item{p.adjust}{If TRUE, p.value adjustment with BH correction is applied
before setting correlations with p.value >= p.val to 0.}

\item{...}{arguments passed to cor.test}
}
\value{
Signature correlation matrix
}
\description{
Pairwise signature correlations from symmetric pivot coordinates
}
\details{
For each pair of signatures correlation based on symmetric pivot
coordinates calculated in individual samples from only positively expressed
signatures. The non-CoDa version of the function is \code{\link{cor_sigs}}.
}
