% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_chr_variants.R, R/generic_functions.R
\name{count_chr_variants}
\alias{count_chr_variants}
\title{Count variants across samples and chromosomes}
\usage{
count_chr_variants(chr_split_grangeslist, perMb = TRUE)

count_chr_variants(chr_split_grangeslist, perMb = TRUE)
}
\arguments{
\item{chr_split_grangeslist}{grangeslist object, result from split_chr
function}

\item{perMb}{if TRUE, variant counts are normalized by chromosome length.
default = TRUE}
}
\value{
counts matrix
}
\description{
Count variants across samples and chromosomes

This function counts number of variants across samples and chromosomes.
The function outputs a matrix with samples as rows and chromosomes as
columns. Counts can be normalized for chromosome lengths with
perMb = TRUE argument.
}
\details{
Provides automated variant counting after variants are split by chromosomes.
Normalization by chromosome length is possible, however, the function does
not accounting for cases when parts of chromosomes are lost.
}
\examples{

SNVs_chr = split_chr(SNVs)
SNVs_chr = lapply( SNVs_chr, function(x) {
   x[["chrY"]] = NULL
   return(x)
}
)
SNV_chr_counts = count_chr_variants(SNVs_chr, perMb = TRUE)

}
