% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{get_chrom_norm_factors}
\alias{get_chrom_norm_factors}
\title{The function splits GRangesList files of samples into list of GRangesList
for individual chromosomes in samples}
\usage{
get_chrom_norm_factors(genome_assembly)
}
\description{
The function splits GRangesList files of samples into list of GRangesList
for individual chromosomes in samples
}
