% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{get_corrupt_regions}
\alias{get_corrupt_regions}
\title{Get regions with of FP cluster variants with given entropy_thresholds
and length of a region with high entropy.}
\usage{
get_corrupt_regions(entropy_vals, threshold, window, step, expand)
}
\description{
Get regions with of FP cluster variants with given entropy_thresholds
and length of a region with high entropy.
}
