% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{get_multinucl_profiles}
\alias{get_multinucl_profiles}
\title{Returns number of multinucleotide substitutions for each sample
and each chromosome from a list of samples with elements as 
GRangesList with individual chromosome variants}
\usage{
get_multinucl_profiles(multisub_chr_list)
}
\description{
Returns number of multinucleotide substitutions for each sample
and each chromosome from a list of samples with elements as 
GRangesList with individual chromosome variants
}
