% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{get_multisub_count_matrix}
\alias{get_multisub_count_matrix}
\title{Generates a count matrix with substitution types as columns
and samples-chrosomosomes as rows}
\usage{
get_multisub_count_matrix(multisub_chr_profile)
}
\description{
Generates a count matrix with substitution types as columns
and samples-chrosomosomes as rows
}
