% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{get_mut_types}
\alias{get_mut_types}
\title{Get mutation types from mutational profiles of sample lists}
\usage{
get_mut_types(SNV_profiles)
}
\description{
Get mutation types from mutational profiles of sample lists
}
