% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{get_variant_entropy}
\alias{get_variant_entropy}
\title{For genomic windows calculate Shannon entropy along with rainfall plots}
\usage{
get_variant_entropy(variants, chr_length, window = 1e+06, step = 5e+05)
}
\description{
For genomic windows calculate Shannon entropy along with rainfall plots
}
