% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor_plot.R
\name{minor_plot}
\alias{minor_plot}
\title{expands the panel with a given factor}
\usage{
minor_plot(ggout, title, expand.factor, expand.factor.y = NULL)
}
\arguments{
\item{ggout}{Input ggplot file}

\item{title}{Title of the output file}

\item{expand.factor}{Expand the output with the given factor}

\item{expand.factor.y}{Expand factor for the y axis if they are different}
}
\description{
expands the panel with a given factor
}
