% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_matrix.R
\name{mut_matrix}
\alias{mut_matrix}
\title{MutationalPatterns mut_matrix function with customizable thread counts}
\usage{
mut_matrix(vcf_list, ref_genome, n_cores)
}
\arguments{
\item{vcf_list}{GRangesList or GRanges object.}

\item{ref_genome}{BSGenome reference genome object}

\item{n_cores}{Number of cores. If not provided, 3/4 of cores will be used.}
}
\value{
96 mutation count matrix
}
\description{
MutationalPatterns mut_matrix function with customizable thread counts
}
\details{
See \link[MutationalPatterns]{mut_matrix}
}
