% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{mut_matrix_chr}
\alias{mut_matrix_chr}
\title{Get mutational matrix in trinucleotide context for individual chromsomes}
\usage{
mut_matrix_chr()
}
\description{
Get mutational matrix in trinucleotide context for individual chromsomes
}
