% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_experiments.R
\name{plot_all_experiments}
\alias{plot_all_experiments}
\title{Function plots repeated estimates of correlation coefficients between
mutational singnatures}
\usage{
plot_all_experiments(all.estimates, title)
}
\arguments{
\item{estimate.list}{List of repeated correlation estimates between signatures}
}
\description{
Function plots repeated estimates of correlation coefficients between
mutational singnatures
}
