% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_plot.R
\name{plot_checkmate_summary}
\alias{plot_checkmate_summary}
\title{Function plots repeated estimates of correlation coefficients between
mutational singnatures}
\usage{
plot_checkmate_summary(estimate.list)
}
\arguments{
\item{estimate.list}{List of repeated correlation estimates between signatures}
}
\description{
Function plots repeated estimates of correlation coefficients between
mutational singnatures
}
