% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R, R/plot_chrwise_counts.R
\name{plot_chrwise_counts}
\alias{plot_chrwise_counts}
\title{Plot chromosome-wise counts with KOs as facets, colored conditions
and grouped for clones}
\usage{
plot_chrwise_counts(chr_wise_counts, KOs, treatment, clones)

plot_chrwise_counts(chr_wise_counts, KOs, treatment, clones)
}
\arguments{
\item{chr_wise_counts}{count matrix, result from count_chr_variants}

\item{KOs}{sample KO's by which plot facets are defined}

\item{treatment}{treatment conditions by which coloring is defined}

\item{clones}{replicates}
}
\description{
Plot chromosome-wise counts with KOs as facets, colored conditions
and grouped for clones

Lineplot variant counts across chromosomes and samples
}
\details{
Plot chromosome-wise counts with samples as facets, colored by conditions,
clones are grouped.
}
