% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R, R/plot_coda_pca.R
\name{plot_coda_pca}
\alias{plot_coda_pca}
\title{Plot CoDa PCA}
\usage{
plot_coda_pca(dt_list, sample_classes, point_size, arrow_length, method)

plot_coda_pca(dt_list, sample_classes, point_size, arrow_length, method)
}
\arguments{
\item{dt_list}{Either list of mutational profiles or a data.frame with
compositional parts(signature intensities, mutation counts) as columns and
samples as rows.}

\item{sample_classes}{A vector of sample classes to be color coded. The
length of this parameter has to be equal to the length of dt_list if list is
provided, or number of rows if data.frame provided.}

\item{point_size}{Point size. default: 3}

\item{arrow_length}{Length of biplot arrows. default: 2}

\item{method}{The method for PCA: classical or robust. default:robust}
}
\description{
Plot CoDa PCA

Classical or robust PCA after CoDa transformation
}
\details{
CoDa transformation is applied on 96-channel or 6-channel mutationa data,
as well as signature composition data before classical or robust PCA.
The matrices have to be positive.
}
