% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R,
%   R/plot_coda_pca_continuous.R
\name{plot_coda_pca_continuous}
\alias{plot_coda_pca_continuous}
\title{Plot CoDa PCA continuous coloring}
\usage{
plot_coda_pca_continuous(
  dt_list,
  continuous,
  palette_name,
  point_size,
  arrow_length,
  method
)

plot_coda_pca_continuous(
  dt_list,
  continuous,
  palette_name,
  point_size,
  arrow_length,
  method
)
}
\arguments{
\item{dt_list}{Either list of mutational profiles or a data.frame with
compositional parts(signature intensities, mutation counts) as columns and
samples as rows.}

\item{continuous}{A vector of annotation values to be color coded. The
length of this parameter has to be equal to the length of dt_list if list is
provided, or number of rows if data.frame provided.}

\item{palette_name}{The palette to be passed to scale_color_distiller function.
default: Spectral.}

\item{point_size}{Point size}

\item{arrow_length}{the length of biplot arrows}

\item{method}{The method for PCA: classical or robust. default:robust}
}
\description{
Plot CoDa PCA continuous coloring

Classical or robust PCA after CoDa transformation for continuous annotations
}
\details{
CoDa transformation is applied on 96-channel or 6-channel mutationa data,
as well as signature composition data before classical or robust PCA with
continuous annotations. The matrices have to be positive.
}
