% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_networks.R
\name{plot_interaction_networks}
\alias{plot_interaction_networks}
\title{Getting a matrix of interactions and returning an interaction network}
\usage{
plot_interaction_networks(
  data_matrix,
  metric_functions,
  filter_list = NULL,
  layout = "stress"
)
}
\arguments{
\item{data_matrix}{input matrix for which to calculate the metrics and
demonstrate with a network}

\item{metric_functions}{is a list with elements as functions which return the
calculated interactions across signatures.}
}
\description{
Getting a matrix of interactions and returning an interaction network
}
\examples{
set.seed(10)
mat = matrix(sample(100, 300, replace = TRUE), 30, 10,
		   dimnames = list(paste0("c", 1:30), letters[1:10]))
mat[, 2] = 2 * mat[,1] + mat[,2]
mat[runif(300) < 0.5] = 0

metric_functions = list(CoDa = cor_coda,
                        cooccurrence = cooccurrence,
                        MI = bcmi,
                        Spearman = function(x) {cor_sigs(x, method = "spearman")})

plot_interaction_networks(mat, metric_functions, 
                                        filter_list = NULL,
                                        layout = "stress")
}
