% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Signature network from adjacency-like matrix}
\usage{
plot_network(
  adjacency_matrix,
  min_threshold = 0.2,
  binary_matrix = FALSE,
  layout,
  expand_limits
)
}
\arguments{
\item{adjacency_matrix}{A sparse matrix with values values indicating the
strength and direction of interaction. Correlations, adjacencies, etc.}

\item{min_threshold}{All values with less absolute value than min_threshold
will be set to 0. Default: 0.2}

\item{binary_matrix}{If binary, the links are constructed from values 1.
Otherwise the matrix is considered a correlation matrix and all 0 values are
set to 0 before network construction. Default: FALSE.}

\item{layout}{Layout passed to ggraph. Default: kk.}

\item{expand_limits}{Argument to control the expansion across x axis.
Default: c(0, 1)}
}
\description{
Signature network from adjacency-like matrix
}
\details{
Network representation of signature interaction matrix. The edges are colored
based on interaction sign and value.
}
