% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R, R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot classical PCA}
\usage{
plot_pca(dt_list, sample_classes, point_size, arrow_length)

plot_pca(dt_list, sample_classes, point_size, arrow_length)
}
\arguments{
\item{dt_list}{Either list of mutational profiles or a data.frame with
compositional parts(signature intensities, mutation counts) as columns and
samples as rows.}

\item{sample_classes}{A vector of sample classes to be color coded. The
length of this parameter has to be equal to the length of dt_list if list is
provided, or number of rows if data.frame provided.}

\item{point_size}{Point size}

\item{arrow_length}{the length of biplot arrows}
}
\description{
Plot classical PCA

Classical PCA without CoDa transformation for comparison
}
\details{
This function is provided solely as a comparison tool agains
\code{\link{plot_coda_pca}} \code{\link{plot_coda_pca_continuous}}.
}
\seealso{
\code{\link{plot_coda_pca}}, \code{\link{plot_coda_pca_continuous}}.
}
