% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot_sub_type_boxplots}
\alias{plot_sub_type_boxplots}
\title{Plot substitution types
### Two types of inputs}
\usage{
plot_sub_type_boxplots(dt_list, sub, sample_classes, treatment)
}
\description{
Plot substitution types
### Two types of inputs
}
