% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{split_by_regions}
\alias{split_by_regions}
\title{Split variants into genomic regions 
exons, introns, utr3, utr5, intergenic}
\usage{
split_by_regions(variants_list, txdb_file, n_cores)
}
\description{
Split variants into genomic regions 
exons, introns, utr3, utr5, intergenic
}
