% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_chr.R
\name{split_chr}
\alias{split_chr}
\title{Split variants into chromosomes}
\usage{
split_chr(vcf_granges_list, n_cores)
}
\arguments{
\item{vcf_granges_list}{GRangesList GRanges variants across samples}

\item{n_cores}{Number of cores. If not provided, 3/4 of cores will be used.}
}
\value{
Returns a list of GRangesLists, for each sample the variants are
split by chromosomes.
}
\description{
Split variants into chromosomes
}
\details{
The function splits GRangesList objects of samples into a list of GRangesList
with variants from individual chromosomes.
}
