% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_calcs.R
\name{summarize_calcs}
\alias{summarize_calcs}
\title{Summary of repeated measurements of signature associations}
\usage{
summarize_calcs(all.estimates)
}
\arguments{
\item{all.estimates}{List of repeated correlation estimates between signatures. Required.}

\item{title}{Plot title. Required.}

\item{rect.lwd}{Line width around rectangles. Default: 0.8.}

\item{col.lims}{Color limits vector to be passed to scale_fill_gradient.
Default: c(-1, 1)}
}
\value{
returns a symmetric matrix of signature interactions, diag is 0.
}
\description{
Over repeated measurements stored in lists of lists, the
function returns 0 if half of the values in each subsampling calculation is 0
for half of replication experiments. Otherwise, the function returns the average
from the experiments where at least half of the values are non-zero.
}
