% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{count_chr_variants}
\alias{count_chr_variants}
\title{This function counts number of variants across samples and chromosomes.
The function outputs a matrix with samples as rows and chromosomes as
columns. Counts can be normalized for chromosome lengths with
chr.norm = TRUE argument.}
\usage{
count_chr_variants(chr_split_grangeslist, chr.norm = TRUE)
}
\description{
import GenomeInfoDb
import parallel detectCores
import parallel mclapply
}
