% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R, R/get_mut_types.R
\name{get_mut_types}
\alias{get_mut_types}
\title{Get mutation types from mutational profiles of sample lists}
\usage{
get_mut_types(SNV_profiles_list, perMb = TRUE, genome_assembly)

get_mut_types(SNV_profiles_list, perMb = TRUE, genome_assembly)
}
\arguments{
\item{SNV_profiles_list}{list of matrices, result from
\code{\link{mut_matrix_chr}} function.}

\item{perMb}{if TRUE, variant counts are normalized by chromosome length.
default = TRUE}
}
\value{
list of matrices with 6-channel substitution type counts for each
sample
}
\description{
Get mutation types from mutational profiles of sample lists

Convert 96 channel mutational profiles to 6
}
\details{
Mutational profiles of samples for individual chromosomes are transformed
into 6 mutation types. Works on a list of matrices.
}
\examples{

SNVs_chr = split_chr(SNVs)
SNV_chr_muts = mut_matrix_chr(SNVs_chr, ref_genome)
SNV_chr_mut_types = get_mut_types(SNV_chr_muts, perMb = TRUE, "hg38")

}
