% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{mut_matrix}
\alias{mut_matrix}
\title{MutationalPatterns mut_matrix added n_cores option}
\usage{
mut_matrix(vcf_list, ref_genome, n_cores)
}
\description{
MutationalPatterns mut_matrix added n_cores option
}
