% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R, R/mut_matrix_chr.R
\name{mut_matrix_chr}
\alias{mut_matrix_chr}
\title{Get mutational matrix in trinucleotide context for individual chromsomes}
\usage{
mut_matrix_chr(chr_split_grangeslist, ref_genome, n_cores)

mut_matrix_chr(chr_split_grangeslist, ref_genome, n_cores)
}
\arguments{
\item{chr_split_grangeslist}{grangeslist object, result from split_chr
function}

\item{ref_genome}{BSGenome reference genome object}

\item{n_cores}{Number of cores. If not provided, 3/4 of cores will be used.}
}
\value{
a list of 96 mutation count matrices
}
\description{
Get mutational matrix in trinucleotide context for individual chromsomes

Get mutational matrix in trinucleotide context for individual chromsomes
}
\details{
Equivalent to mut_matrix from MutationalPatterns, works on a list of
grangeslist with chr-variants.
}
