% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_experiments.R
\name{plot_all_experiments}
\alias{plot_all_experiments}
\title{Plots repeated estimates of correlation coefficients between mutational
signatures}
\usage{
plot_all_experiments(all.estimates, title, rect.lwd, col.lims)
}
\arguments{
\item{title}{Plot title. Required.}

\item{rect.lwd}{Line width around rectangles. Default: 0.8.}

\item{col.lims}{Color limits vector to be passed to scale_fill_gradient.
Default: c(-1, 1)}

\item{estimate.list}{List of repeated correlation estimates between signatures. Required.}
}
\description{
Plots repeated estimates of correlation coefficients between mutational
signatures
}
