% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot_chrwise_counts}
\alias{plot_chrwise_counts}
\title{Plot chromosome-wise counts with KOs as facets, colored conditions
and grouped for clones}
\usage{
plot_chrwise_counts(chr_wise_counts, KOs, treatment, clones)
}
\description{
Plot chromosome-wise counts with KOs as facets, colored conditions
and grouped for clones
}
