% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot_coda_pca_continuous}
\alias{plot_coda_pca_continuous}
\title{Plot CoDa PCA continuous coloring}
\usage{
plot_coda_pca_continuous(dt_list, continuous, palette_name)
}
\description{
Plot CoDa PCA continuous coloring
}
