% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R, R/split_by_regions.R
\name{split_by_regions}
\alias{split_by_regions}
\title{Split variants into genomic regions 
exons, introns, utr3, utr5, intergenic}
\usage{
split_by_regions(variants_list, txdb_file, n_cores)

split_by_regions(variants_list, txdb_file, n_cores)
}
\arguments{
\item{variants_list}{List of GRanges objects. Each element of a list is the
set of sample variants.}

\item{txdb_file}{TxDb object matching with the genome assembly of variants}

\item{n_cores}{Number of cores. If not provided, 3/4 of cores will be used.#'}
}
\value{
A list of genomic regions (exons, introns, etc.), containing a list of
samples with variants residing in corresponding regions.
}
\description{
Split variants into genomic regions 
exons, introns, utr3, utr5, intergenic

Split variants into genomic regions: exons, introns, utr3, utr5, intergenic
}
\details{
Variants are split based on genomic regions - exons, introns, 3'UTR,
5'UTR, transcripts, intergenic regions defined from TxDb.
}
\examples{
SNVs_hg38_split = split_by_regions(SNVs, TxDb.Hsapiens.UCSC.hg38.knownGene)

}
