% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{split_chr}
\alias{split_chr}
\title{The function splits GRangesList files of samples into list of GRangesList
for individual chromosomes in samples}
\usage{
split_chr(vcf_granges_list, n_cores)
}
\description{
import GenomeInfoDb
import parallel detectCores
import parallel mclapply
import GenomicFeatures getChromInfoFromUCSC
import data.table
}
