% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitVCFToMutClasses.R
\name{splitVCFToMutClasses}
\alias{splitVCFToMutClasses}
\title{A function splitting vcf files into SNVs, multinucleotide substitutions and indels}
\usage{
splitVCFToMutClasses(vcffile, n_cores)
}
\description{
import GenomeInfoDb
import parallel detectCores
import parallel mclapply
import dplyr bind_rows
}
\examples{
splitVCFtoMutClasses(vcf_granges)

}
