% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitVCFToMutClasses.R
\name{splitVCFToMutClasses}
\alias{splitVCFToMutClasses}
\title{A function splitting vcf files into SNVs, multinucleotide substitutions and indels}
\usage{
splitVCFToMutClasses(vcffile)
}
\description{
A function splitting vcf files into SNVs, multinucleotide substitutions and indels
}
\examples{
splitVCFtoMutClasses(vcf_granges)

}
