% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_pivot_coords.R
\name{symmetric_pivot_coords}
\alias{symmetric_pivot_coords}
\title{A function returning the symmetric pivots for i-th and j-th cols}
\usage{
symmetric_pivot_coords(x, i, j, rand.add = FALSE, ...)
}
\arguments{
\item{x}{Data.frame or a matrix of signature intensities(or counts).
Signatures need to correspond to columns.}

\item{i}{Index of a signature for which the pivots should be calculated}

\item{j}{Same as i}

\item{rand.add}{if random noise was added to data, then rand.add = TRUE,
otherwise rand.add = FALSE. If FALSE balZavout is calculated similar to
coda_novel. Default: FALSE.}
}
\value{
A matrix, where the first two cols are the
symmetric coordinates
}
\description{
A function returning the symmetric pivots for i-th and j-th cols
}
